requirements.txt
Django == 2.2.3
django-tinymce == 2.8.0
django-bootstrap4 == 0.0.8
singlemodeladmin == 0.9
django-compressor == 2.3
Pillow == 6.1.0
feedparser == 5.2.1
django-admin-sortable == 2.1.17
django-admin-tools == 0.8.1
jsonschema == 3.0.2